[branch-2.9] Support mock final class with `PowerMockito` (#17327)
